notebooks/South-Africa-Field-Boundary/requirements.txt
radiant-mlhub==0.1.*
pandas==1.*
torch
torchvision
tensorflow==2.4.0
keras==2.3.1
segmentation_models
matplotlib==3.*
scikit-image==0.*
scikit-learn
numpy
rasterio
ipywidgets
Pillow
h5py==2.10.0
folium
opencv-python
rioxarray
tqdm
pycocotools
seaborn
shapely
